# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cumments API
  version: 0.25.4
  description: |
    Backend-only comment API backed by Matrix. Writes are asynchronous:
    POST/PATCH/DELETE accept a submission and return `202 { "submission_id" }`;
    the projected comment later carries the same `submission_id`.

    Comment reads use the RFC 10008 `QUERY` method. OpenAPI does not support
    custom verbs, so each QUERY operation is represented with the
    `x-cumments-query` extension and is sent as `QUERY` on the wire.
servers:
  - url: http://localhost:7931
...
security: []
paths:
...
  /api/v1/operator/rooms/{room_id}/upgrade:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [Operator]
      summary: Upgrade a comment room (operator)
      operationId: operatorUpgradeRoom
      description: |
        Operator mirror of the site-level upgrade endpoint: upgrades a
        registered active comment room and converges the replacement. The
        target version must be newer than the room's current version; the
        upgrade is executed by the AS bot, which stays the replacement
        room's creator.
      security:
        - operatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeRoomRequest"
      responses:
        "200":
          description: Replacement room converged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRoomResponse"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/operator/rooms/{room_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    delete:
      tags: [Operator]
      summary: Retire a comment room (operator)
      operationId: operatorRetireRoom
      description: |
        Operator mirror of the site-level page retirement: marks the
        registered active room `Retired`, stopping new writes, and the
        background reconciler leaves the Matrix room (rename, alias removal,
        AS sender and virtual users) and clears local projections.
      security:
        - operatorToken: []
      responses:
        "200":
          description: Room marked retiring
          content:
            application/json:
              schema:
                type: object
                required: [room_id, status]
                properties:
                  room_id:
                    type: string
                  status:
                    type: string
                    enum: [retiring]
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
    claimToken:
      type: apiKey
      in: header
      name: X-Cumments-Claim-Token
  parameters:
...
    RoomId:
      name: room_id
      in: path
      required: true
      schema:
        type: string
        description: Matrix room ID, percent-encoded
...
  responses:
    Error:
      description: RFC 9457 problem details response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri-reference
          description: Canonical URI of the problem type
        title:
          type: string
          description: Stable short human-readable name of the type
        status:
          type: integer
          format: int32
          description: HTTP status code, identical to the response status
        detail:
          type: string
          description: Occurrence-specific explanation for the client
        code:
          type: string
          description: Stable machine-readable slug; the fragment (after `#`) of `type`
        details:
          type: object
          nullable: true
...
    UpgradeRoomRequest:
      type: object
      required: [new_version]
      properties:
        new_version:
          type: string
          description: Target Matrix room version; must be newer than the room's current version.
...
    UpgradeRoomResponse:
      type: object
      required: [room_id, new_version, replacement_room]
      properties:
        room_id:
          type: string
        new_version:
          type: string
        replacement_room:
          type: string
...